+++ source/rafcon/gui/mygaphas/items/state.py
@@ -59,13 +59,15 @@ class StateView(Element):
 
     _map_handles_port_v = {}
 
-    def __init__(self, state_m, size, hierarchy_level):
+    def __init__(self, state_m, size, custom_background_color, background_color, hierarchy_level):
         super(StateView, self).__init__(size[0], size[1])
         assert isinstance(state_m, AbstractStateModel)
         # Reapply size, as Gaphas sets default minimum size to 1, which is too large for highly nested states
         self.min_width = self.min_height = 0
         self.width = size[0]
         self.height = size[1]
+        self.custom_background_color = custom_background_color
+        self.background_color = background_color
 
         self._c_min_w = self._constraints[0]
         self._c_min_h = self._constraints[1]
@@ -351,6 +353,8 @@ class StateView(Element):
         self.position = state_meta['rel_pos']
         self.width = state_meta['size'][0]
         self.height = state_meta['size'][1]
+        self.custom_background_color = state_meta['custom_background_color']
+        self.background_color = state_meta['background_color']
         self.update_minimum_size_of_children()
 
         def update_port_position(port_v, meta_data):
@@ -450,7 +454,10 @@ class StateView(Element):
             if not context.draw_all:
                 inner_nw, inner_se = self.get_state_drawing_area(self)
                 c.rectangle(inner_nw.x, inner_nw.y, inner_se.x - inner_nw.x, inner_se.y - inner_nw.y)
-                c.set_source_rgba(*get_col_rgba(state_background_color))
+                if self.custom_background_color:
+                    c.set_source_rgba(self.background_color[0], self.background_color[1], self.background_color[2], 0.5)
+                else:
+                    c.set_source_rgba(*get_col_rgba(state_background_color))
                 c.fill_preserve()
                 c.set_source_rgba(*get_col_rgba(state_border_outline_color, self.transparency))
                 c.set_line_width(default_line_width)